Brings in ~548 commits of upstream work, including the video container
support (ContainerMTS/MP4/MPG, DemuxerMTS, PacedVideoOutput,
VideoPlayer) needed to eventually mux audio+video into a single MPEG-TS
RTSP stream.

Conflicts resolved:
- .gitignore: took main's broader `.*`/`.*/` patterns (already covers
  build_check/, .vscode/, .DS_Store, etc. that rtsp-video listed
  individually)
- tests-cmake/CMakeLists.txt: took main's restructured version (single
  add_subdirectory per category); re-added the new rtsp-video test by
  wiring it into tests-cmake/network/CMakeLists.txt instead, alongside
  main's own rtsp/url-test/hdlc entries there

Verified after merge: tests-cmake configures cleanly and the rtsp /
rtsp-video targets both still build against the merged RTSP headers.

+++ src/AudioTools/Concurrency/RTOS/QueueRTOS.h
@@ -17,6 +17,9 @@ namespace audio_tools {
 /**
  * @brief FIFO Queue whch is based on the FreeRTOS queue API.
  * The default allocator will allocate the memory from psram if available
+ * 
+ * @note Supported by all FreeRTOS platforms
+ * 
  * @ingroup collections
  * @ingroup concurrency
  * @author Phil Schatzmann
@@ -47,7 +50,7 @@ class QueueRTOS {
   void setWriteMaxWait(TickType_t ticks) { write_max_wait = ticks; }
 
   /// (Re-)defines the size
-  bool resize(int size) {
+  bool resize(size_t size) {
     bool result = true;
     TRACED();
     if (size != queue_size) {
